services/core/requirements.txt
# Core Platform Service - Aarya Clothing
# User Management, Authentication, and Session Services

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
